BIG update to scripting, as it now has enough features to function well as a language. Added text boxes and the ability to have multiple of them. General improvements to the codebase.

+++ Engine/src/openGL/OpenGL.h
@@ -1,10 +1,9 @@
 #pragma once
-#include "src\openGL\compile\StandardOpenGL.h"
 
 void initOpenGL();
 void drawScene();
 
-float startFrame();
+double startFrame();
 void endFrame();
 
 void setWireframe(bool isOn);
